fix saving dataset to HF, toolcall sanitizing
